% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seapiperdata.R
\name{custom_yaml_to_seapiperdata}
\alias{custom_yaml_to_seapiperdata}
\title{Build seaPiperData from a YAML file of RDS paths}
\usage{
custom_yaml_to_seapiperdata(...)
}
\arguments{
\item{...}{arguments passed to \code{seapiperdata_from_yaml()}}
}
\description{
Deprecated alias of \code{seapiperdata_from_yaml()}.
}
