% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seapiperdata.R
\name{make_seapiperdata}
\alias{make_seapiperdata}
\title{Build seaPiperData from basic inputs}
\usage{
make_seapiperdata(...)
}
\arguments{
\item{...}{arguments passed to \code{seapiperdata_from_objects()}}
}
\description{
Deprecated alias of \code{seapiperdata_from_objects()}.
}
