% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seapiperdata.R
\name{rseasnap_to_seapiperdata}
\alias{rseasnap_to_seapiperdata}
\title{Convert Rseasnap pipeline output into seaPiperData}
\usage{
rseasnap_to_seapiperdata(...)
}
\arguments{
\item{...}{arguments passed to \code{seapiperdata_from_rseasnap()}}
}
\description{
Deprecated alias of \code{seapiperdata_from_rseasnap()}.
}
