% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{seapiper}
\alias{seapiper}
\title{seaPiper Workflow output explorer}
\usage{
seapiper(data, title = "Workflow output explorer", debug_panel = FALSE)
}
\arguments{
\item{data}{a \code{seaPiperData} object created by \code{seapiperdata_from_rseasnap},
\code{seapiperdata_from_objects}, or \code{seapiperdata_from_yaml}}

\item{title}{Name of the pipeline to display}

\item{debug_panel}{show a debugging panel}
}
\description{
seaPiper Workflow output explorer for the sea-snap pipeline
}
\details{
Launch a shiny app for viewing / exploring the results of a differential
expression analysis performed in sea-snap.

Launching is much faster if the large objects (contrasts, tmod
databases) do not need to be loaded each time the pipeline browser is
started.
}
\examples{
if(interactive()) {
  example_dir <- system.file("extdata/example_pipeline", package="Rseasnap")
  conf_f      <- file.path(example_dir, "DE_config.yaml")
  pip         <- load_de_pipeline(config_file = conf_f)
  data        <- seapiperdata_from_rseasnap(pip)
  seapiper(data)
}
}
