% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seapiperdata.R
\name{seapiperdata_from_rseasnap}
\alias{seapiperdata_from_rseasnap}
\title{Convert Rseasnap pipeline output into seaPiperData}
\usage{
seapiperdata_from_rseasnap(
  pip,
  primary_id = "PrimaryID",
  annot = NULL,
  cntr = NULL,
  tmod_res = NULL,
  tmod_dbs = NULL
)
}
\arguments{
\item{pip}{pipeline returned by \code{load_de_pipeline}}

\item{primary_id}{name of the column in the annotion data frame
which corresponds to the primary gene identifier (including the
row names of the contrasts results in the cntr object)}

\item{annot}{annotation returned by \code{get_annot}}

\item{cntr}{contrasts list returned by \code{get_contrasts}}

\item{tmod_res}{tmod results returned by \code{get_tmod_res}}

\item{tmod_dbs}{tmod databases returned by \code{get_tmod_dbs}}
}
\description{
Prepare seaPiper-ready data from a Rseasnap pipeline (or list of pipelines).
}
